Rename {inner,outer}-PEC.js for improved clarity.
