+++ runtime/arc.js
@@ -12,7 +12,7 @@
 import {assert} from '../platform/assert-web.js';
 import {Type} from './type.js';
 import {handleFor} from './handle.js';
-import {OuterPEC} from './outer-PEC.js';
+import {OuterPEC} from './particle-execution-host.js';
 import {Recipe} from './recipe/recipe.js';
 import {Manifest} from './manifest.js';
 import {Description} from './description.js';
