--- runtime/inner-PEC.js
@@ -1,248 +1,0 @@
-/**
- * @license
- * Copyright (c) 2017 Google Inc. All rights reserved.
- * This code may only be used under the BSD style license found at
- * http://polymer.github.io/LICENSE.txt
- * Code distributed by Google as part of this project is also
- * subject to an additional IP rights grant found at
- * http://polymer.github.io/PATENTS.txt
- */
-'use strict';
-
-import {handleFor} from './handle.js';
-import {assert} from '../platform/assert-web.js';
-import {PECInnerPort} from './api-channel.js';
-import {StorageProxy} from './storage-proxy.js';
-
-export class InnerPEC {
-  constructor(port, idBase, loader) {
-    this._apiPort = new PECInnerPort(port);
-    this._particles = [];
-    this._idBase = idBase;
-    this._nextLocalID = 0;
-    this._loader = loader;
-    this._pendingLoads = [];
-
-    /*
-     * This code ensures that the relevant types are known
-     * in the scope object, because otherwise we can't do
-     * particleSpec resolution, which is currently a necessary
-     * part of particle construction.
-     *
-     * Possibly we should eventually consider having particle
-     * specifications separated from particle classes - and
-     * only keeping type information on the arc side.
-     */
-    this._apiPort.onDefineHandle = ({type, identifier, name}) => {
-      return new StorageProxy(identifier, type, this._apiPort, this, name);
-    };
-
-    this._apiPort.onCreateHandleCallback = ({type, id, name, callback}) => {
-      let proxy = new StorageProxy(id, type, this._apiPort, this, name);
-      return [proxy, () => callback(proxy)];
-    };
-
-    this._apiPort.onMapHandleCallback = ({id, callback}) => {
-      return [id, () => callback(id)];
-    };
-
-    this._apiPort.onCreateSlotCallback = ({hostedSlotId, callback}) => {
-      return [hostedSlotId, () => callback(hostedSlotId)];
-    };
-
-    this._apiPort.onInnerArcRender = ({transformationParticle, transformationSlotName, hostedSlotId, content}) => {
-      transformationParticle.renderHostedSlot(transformationSlotName, hostedSlotId, content);
-    };
-
-    this._apiPort.onStop = () => {
-      if (global.close) {
-        global.close();
-      }
-    };
-
-    this._apiPort.onInstantiateParticle =
-      ({id, spec, handles}) => this._instantiateParticle(id, spec, handles);
-
-    this._apiPort.onSimpleCallback = ({callback, data}) => callback(data);
-
-    this._apiPort.onConstructArcCallback = ({callback, arc}) => callback(arc);
-
-    this._apiPort.onAwaitIdle = ({version}) =>
-      this.idle.then(a => {
-        // TODO: dom-particles update is async, this is a workaround to allow dom-particles to
-        // update relevance, after handles are updated. Needs better idle signal.
-        setTimeout(() => { this._apiPort.Idle({version, relevance: this.relevance}); }, 0);
-      });
-
-    this._apiPort.onUIEvent = ({particle, slotName, event}) => particle.fireEvent(slotName, event);
-
-    this._apiPort.onStartRender = ({particle, slotName, contentTypes}) => {
-      /** @class Slot
-       * A representation of a consumed slot. Retrieved from a particle using
-       * particle.getSlot(name)
-       */
-      class Slotlet {
-        constructor(pec, particle, slotName) {
-          this._slotName = slotName;
-          this._particle = particle;
-          this._handlers = new Map();
-          this._pec = pec;
-          this._requestedContentTypes = new Set();
-        }
-        get particle() { return this._particle; }
-        get slotName() { return this._slotName; }
-        get isRendered() { return this._isRendered; }
-        /** @method render(content)
-         * renders content to the slot.
-         */
-        render(content) {
-          this._pec._apiPort.Render({particle, slotName, content});
-
-          Object.keys(content).forEach(key => { this._requestedContentTypes.delete(key); });
-          // Slot is considered rendered, if a non-empty content was sent and all requested content types were fullfilled.
-          this._isRendered = this._requestedContentTypes.size == 0 && (Object.keys(content).length > 0);
-        }
-        /** @method registerEventHandler(name, f)
-         * registers a callback to be invoked when 'name' event happens.
-         */
-        registerEventHandler(name, f) {
-          if (!this._handlers.has(name)) {
-            this._handlers.set(name, []);
-          }
-          this._handlers.get(name).push(f);
-        }
-        clearEventHandlers(name) {
-          this._handlers.set(name, []);
-        }
-        fireEvent(event) {
-          for (let handler of this._handlers.get(event.handler) || []) {
-            handler(event);
-          }
-        }
-      }
-
-      particle._slotByName.set(slotName, new Slotlet(this, particle, slotName));
-      particle.renderSlot(slotName, contentTypes);
-    };
-
-    this._apiPort.onStopRender = ({particle, slotName}) => {
-      assert(particle._slotByName.has(slotName),
-        `Stop render called for particle ${particle.name} slot ${slotName} without start render being called.`);
-      particle._slotByName.delete(slotName);
-    };
-  }
-
-  generateIDComponents() {
-    return {base: this._idBase, component: () => this._nextLocalID++};
-  }
-
-  generateID() {
-    return `${this._idBase}:${this._nextLocalID++}`;
-  }
-
-  innerArcHandle(arcId, particleId) {
-    let pec = this;
-    return {
-      createHandle: function(type, name) {
-        return new Promise((resolve, reject) =>
-          pec._apiPort.ArcCreateHandle({arc: arcId, type, name, callback: proxy => {
-            let h = handleFor(proxy, proxy.type.isCollection, name, particleId);
-            h.entityClass = (proxy.type.isCollection ? proxy.type.primitiveType() : proxy.type).entitySchema.entityClass();
-            resolve(h);
-          }}));
-      },
-      mapHandle: function(handle) {
-        return new Promise((resolve, reject) =>
-          pec._apiPort.ArcMapHandle({arc: arcId, handle, callback: id => {
-            resolve(id);
-          }}));
-      },
-      createSlot: function(transformationParticle, transformationSlotName, hostedParticleName, hostedSlotName) {
-        return new Promise((resolve, reject) =>
-          pec._apiPort.ArcCreateSlot({arc: arcId, transformationParticle, transformationSlotName, hostedParticleName, hostedSlotName, callback: hostedSlotId => {
-            resolve(hostedSlotId);
-          }}));
-      },
-      loadRecipe: function(recipe) {
-        // TODO: do we want to return a promise on completion?
-        return new Promise((resolve, reject) =>
-          pec._apiPort.ArcLoadRecipe({arc: arcId, recipe, callback: a => {
-            if (a == undefined)
-              resolve();
-            else
-              reject(a);
-          }}));
-      }
-    };
-  }
-
-  defaultCapabilitySet() {
-    return {
-      constructInnerArc: particle => {
-        return new Promise((resolve, reject) =>
-          this._apiPort.ConstructInnerArc({callback: arcId => {resolve(this.innerArcHandle(arcId, particle.id));}, particle}));
-      }
-    };
-  }
-
-  async _instantiateParticle(id, spec, proxies) {
-    let name = spec.name;
-    let resolve = null;
-    let p = new Promise((res, rej) => resolve = res);
-    this._pendingLoads.push(p);
-    let clazz = await this._loader.loadParticleClass(spec);
-    let capabilities = this.defaultCapabilitySet();
-    let particle = new clazz(); // TODO: how can i add an argument to DomParticle ctor?
-    particle.id = id;
-    particle.capabilities = capabilities;
-    this._particles.push(particle);
-
-    let handleMap = new Map();
-    let registerList = [];
-    proxies.forEach((proxy, name) => {
-      let connSpec = spec.connectionMap.get(name);
-      let handle = handleFor(proxy, proxy.type.isCollection, name, id, connSpec.isInput, connSpec.isOutput);
-      let type = proxy.type.isCollection ? proxy.type.primitiveType() : proxy.type;
-      if (type.isEntity) {
-        handle.entityClass = type.entitySchema.entityClass();
-      }
-      handleMap.set(name, handle);
-
-      // Defer registration of handles with proxies until after particles have a chance to
-      // configure them in setHandles.
-      registerList.push({proxy, particle, handle});
-    });
-
-    return [particle, async () => {
-      resolve();
-      let idx = this._pendingLoads.indexOf(p);
-      this._pendingLoads.splice(idx, 1);
-      await particle.setHandles(handleMap);
-      registerList.forEach(({proxy, particle, handle}) => proxy.register(particle, handle));
-    }];
-  }
-
-  get relevance() {
-    let rMap = new Map();
-    this._particles.forEach(p => {
-      if (p.relevances.length == 0)
-        return;
-      rMap.set(p, p.relevances);
-      p.relevances = [];
-    });
-    return rMap;
-  }
-
-  get busy() {
-    if (this._pendingLoads.length > 0)
-      return true;
-    return false;
-  }
-
-  get idle() {
-    if (!this.busy) {
-      return Promise.resolve();
-    }
-    return Promise.all(this._pendingLoads.concat(this._particles.map(particle => particle.idle))).then(() => this.idle);
-  }
-}
