+++ runtime/test/debug/outer-port-attachments-tests.js
@@ -13,7 +13,7 @@ import {assert} from '../chai-web.js';
 import * as util from '../../testing/test-util.js';
 import {Arc} from '../../arc.js';
 import {MessageChannel} from '../../message-channel.js';
-import {InnerPEC} from '../../inner-PEC.js';
+import {InnerPEC} from '../../particle-execution-context.js';
 import {StubLoader} from '../../testing/stub-loader.js';
 import {DevtoolsForTests} from '../../debug/devtools-connection.js';
 import {Random} from '../../random.js';
