+++ runtime/test/particle-shape-loading-test.js
@@ -13,7 +13,7 @@ import {assert} from './chai-web.js';
 import * as util from '../testing/test-util.js';
 import {Arc} from '../arc.js';
 import {MessageChannel} from '../message-channel.js';
-import {InnerPEC} from '../inner-PEC.js';
+import {InnerPEC} from '../particle-execution-context.js';
 import {Loader} from '../loader.js';
 import {StubLoader} from '../testing/stub-loader.js';
 import {Recipe} from '../recipe/recipe.js';
