+++ shell/source/worker-entry.js
@@ -6,7 +6,7 @@
 // subject to an additional IP rights grant found at
 // http://polymer.github.io/PATENTS.txt
 
-import {InnerPEC} from '../../runtime/inner-PEC.js';
+import {InnerPEC} from '../../runtime/particle-execution-context.js';
 import {BrowserLoader} from './browser-loader.js';
 
 const log = console.log.bind(console, `%cworker-entry`, `background: #12005e; color: white; padding: 1px 6px 2px 7px; border-radius: 6px;`);
